docs: Tweaks to the platform setup docs
I've added some additional tweaks to the documents for pieces
that were not 100% clear for me during creation of a platform.

The hard-coding of the profile I've replaced as an environment
variable, AWS_PROFILE, I am aware that the aws cli will use this
if it is set however I feel that explicitly declaring it make
its use more clear.

Changed the base64 decode to use --decode as this is compatible
across distributions however -d is not.

Changes some of the docs around the vault setup as it currently
only comes up on specific machines in the cluster and being able
to identify which ones to run the vault commands against is useful
event if it happens to be the local machines

I've also added a small bit of info at the point of setting up the
authz/authn files for the kubernetes cluster which are then encrypted.