Increase TCP timeout for docker workers elb
AWS ELB has a TCP connection timeout of 60 seconds, which means if there
is no activity (no output, etc) coming from docker host via an ELB, the
ELB terminates the connection. And the way drone behaves is a bit
fragile, because as soon as it sees a connection closed, it starts to
terminate containers and to clean up.